refactor(server): Split event filter into separate file
