+++ src/server/ua_server_internal.h
@@ -387,11 +387,20 @@ void monitoredItem_sampleCallback(UA_Server *server, UA_MonitoredItem *monitored
 UA_Subscription *
 UA_Server_getSubscriptionById(UA_Server *server, UA_UInt32 subscriptionId);
 
+#ifdef UA_ENABLE_SUBSCRIPTIONS_EVENTS
 UA_StatusCode
 triggerEvent(UA_Server *server, const UA_NodeId eventNodeId,
              const UA_NodeId origin, UA_ByteString *outEventId,
              const UA_Boolean deleteEventNode);
 
+/* Filters the given event with the given filter and writes the results into a
+ * notification */
+UA_StatusCode
+filterEvent(UA_Server *server, UA_Session *session,
+            const UA_NodeId *eventNode, UA_EventFilter *filter,
+            UA_EventFieldList *efl, UA_EventFilterResult *result);
+
+#endif /* UA_ENABLE_SUBSCRIPTIONS_EVENTS */
 #endif /* UA_ENABLE_SUBSCRIPTIONS */
 
 UA_BrowsePathResult
